% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_priors.R
\name{add_priors}
\alias{add_priors}
\title{Add priors for sizedist models}
\usage{
add_priors(pars, Z_sd = 10)
}
\arguments{
\item{pars}{pars object}

\item{Z_sd}{standard deviation for Z}
}
\value{
list of priors appended to pars list
}
\description{
Add priors for sizedist models
}
