% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{add_sampling_noise}
\alias{add_sampling_noise}
\title{Simulate noise for any variable from a distribution with mean = 0}
\usage{
add_sampling_noise(data, var, sd = 0.1, overwrite = TRUE)
}
\arguments{
\item{data}{simulated dataframe or own data}

\item{var}{variable you wish to add noise to}

\item{sd}{standard deviation}
}
\value{
dataframe where variable w
}
\description{
Simulate noise for any variable from a distribution with mean = 0
}
