% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_bin_data.R
\name{compose_bin_data}
\alias{compose_bin_data}
\title{Compile data frame into Stan friendly format for size distribution models}
\usage{
compose_bin_data(data, ...)
}
\arguments{
\item{data}{A data frame containing age and/or size bins counts created using \code{add_bins()} and \code{create_age_bin_counts()}}

\item{...}{Arguments passed to tidybayes::compose_data}
}
\value{
A data list where each list element is a column from the original data frame
}
\description{
Compile data frame into Stan friendly format for size distribution models
}
\seealso{
\code{fit_age()} and \code{create_age_bin_counts()}
}
