% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_data.R
\name{compose_growth_data}
\alias{compose_growth_data}
\title{Compile age-size data into Stan friendly format for growth models}
\usage{
compose_growth_data(data, age_var, size_var)
}
\arguments{
\item{data}{A data frame containing age and/or size bins counts created using \code{summarise_bin_counts}}

\item{age_var}{Name of age variable}

\item{size_var}{Name of size variable}
}
\value{

}
\description{
Compile age-size data into Stan friendly format for growth models
}
\examples{
\dontrun{
data <- simulate_population(pars = pars)
compose_growth_data(data, age, size)
}
}
