% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_age_bin_counts.R
\name{create_age_bin_counts}
\alias{create_age_bin_counts}
\title{Creates bins}
\usage{
create_age_bin_counts(data, bin_width_age = 1)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. tibble) containing \code{age} as a column}

\item{bin_width_age}{The bin width for age, usually the smallest resolution of re-ssampling e.g. 1 day}
}
\value{
a data frame containing lower and upper bounds as well as counts for each \code{age_bin}
}
\description{
Creates bins
}
\examples{
\dontrun{
fishlarvae \%>\% add_bins() \%>\% create_age_bin_counts()
}

}
\seealso{
\code{simulate_population()} \code{add_bins()}
}
