% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bin_counts.R
\name{create_size_bin_counts}
\alias{create_size_bin_counts}
\title{Counts for size bin data}
\usage{
create_size_bin_counts(data, bin_width = 0.1)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. tibble) containing \code{age} as a column}

\item{bin_width}{The bin width for age, usually the smallest resolution of re-sampling e.g. 1 day}
}
\value{
a data frame containing lower and upper bounds as well as counts for each \code{size_bin}
}
\description{
Counts for size bin data
}
\seealso{
\code{simulate_population()} \code{add_bins()}
}
