% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R, R/simulate_population.R
\name{default_pars_model1}
\alias{default_pars_model1}
\alias{default_pars}
\title{Default parameters for size-distribution models}
\usage{
default_pars_model1()

default_pars(model)
}
\arguments{
\item{model}{name of size-distribution model being modeled}
}
\value{
list of parameter values for the size-distribution model, including model name
}
\description{
Default parameters for size-distribution models
}
