% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_priors.R
\name{default_priors}
\alias{default_priors}
\alias{default_priors_model1a}
\alias{default_priors_model1b}
\alias{default_priors_model1c}
\title{Load default priors for a specific model}
\usage{
default_priors(pars, model = pars$model)

default_priors_model1a()

default_priors_model1b()

default_priors_model1c()
}
\arguments{
\item{pars}{pars list object you want to join priors to}

\item{model}{name of the model, takes value in pars$model}
}
\value{
list object with original pars with priors joined at the end
}
\description{
Load default priors for a specific model
}
\examples{
pars <- default_pars("model1")
pars <- pars \%>\% purrr::list_modify(model = "model1a")
pars_new <- default_priors("model1a", pars)
}
