% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size_dist.R
\name{extract_fitted}
\alias{extract_fitted}
\alias{extract_fitted_model1a}
\alias{extract_fitted_model1b}
\alias{extract_fitted_model1c}
\alias{extract_fitted_model1d}
\title{Extract fitted values for Z, R, g}
\usage{
extract_fitted(fit, pars)

extract_fitted_model1a(fit, pars)

extract_fitted_model1b(fit, pars)

extract_fitted_model1c(fit, pars)

extract_fitted_model1d(fit, pars)
}
\arguments{
\item{fit}{model fit for growth model}

\item{pars}{pars list used for model fitting}
}
\value{
list containing fitted values predicted by the model
}
\description{
Extract fitted values for Z, R, g
}
