% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_age.R
\name{fit_age}
\alias{fit_age}
\title{Simple Mortality Age Model with Stan}
\usage{
fit_age(data, ...)
}
\arguments{
\item{data}{Stan-friendly data list created by XX function}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Simple Mortality Age Model with Stan
}
