% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sizedist.R
\name{fit_sizedist}
\alias{fit_sizedist}
\title{Fit size distribution model with Stan}
\usage{
fit_sizedist(data, ...)
}
\arguments{
\item{data}{Stan-friendly data list created by \code{compose_data} function}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit size distribution model with Stan
}
