% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pars.R
\name{join_pars}
\alias{join_pars}
\title{Join parameters to Stan-composed data}
\usage{
join_pars(data, pars, var)
}
\arguments{
\item{data}{Data in Stan friendly format}

\item{pars}{model pars e.g. default_model_pars("model1")}

\item{var}{Name strings}
}
\value{
List with variables from pars appended
}
\description{
Join parameters to Stan-composed data
}
