% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pars.R
\name{keep_fit_pars_model}
\alias{keep_fit_pars_model}
\title{Function to prune down hyper-pars}
\usage{
keep_fit_pars_model(pars)
}
\arguments{
\item{pars}{list object containing model$}
}
\description{
Function to prune down hyper-pars
}
