% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_pars.R
\name{mutate_pars}
\alias{mutate_pars}
\title{Create, modify and remove model parameters}
\usage{
mutate_pars(pars, ...)
}
\arguments{
\item{pars}{The name of the model in the format default_pars("model1")}

\item{...}{Name and value pair:
\itemize{
\item Set value to NULL to delete parameter
\item Values that are character strings need to be in quotes e.g. name of the model
}}
}
\description{
Create, modify and remove model parameters
}
