% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size_dist.R
\name{plot_size_dist}
\alias{plot_size_dist}
\title{Plot the size distribution}
\usage{
plot_size_dist(data, pars, binwidth, fitted = NULL)
}
\arguments{
\item{data}{a dataframe with a simulated size, size_sampled}

\item{pars}{list of parameter values for the size-distribution model, including model name}

\item{binwidth}{The bin width for age, usually the smallest resolution of resampling e.g. 1 day}

\item{fitted}{is a list containing estimated parameters from a stan mortality_size model.}
}
\value{
a ggplot of size-frequency histogram with fitted function line
}
\description{
This function is for plotting the size distribution of simulated populations.
}
