% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{sample_birth_times}
\alias{sample_birth_times}
\title{Samples time of birth from a homogeneous Poisson process.}
\usage{
sample_birth_times(R, time_end)
}
\arguments{
\item{R}{average birth rate per unit time}

\item{time_end}{Duration of simulation}
}
\description{
Samples are drawn assuming a homogeneous Poisson process. Under this distribution, the inter-arrival times follow a log distribution. https://en.wikipedia.org/wiki/Poisson_distribution
}
