% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{sample_individual_variation}
\alias{sample_individual_variation}
\title{Sample variation in parameters among individuals across the population under a particular size-distribution model}
\usage{
sample_individual_variation(n, pars)
}
\arguments{
\item{n}{number of individuals to sample}

\item{pars}{list of parameter values for the size-distribution model, including model name}
}
\value{
list of parameter values for the size-distribution model, including model name


}
\description{
Sample variation in parameters among individuals across the population under a particular size-distribution model
}
