% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_catch_data.R
\name{simulate_catch_data}
\alias{simulate_catch_data}
\title{Simulating a dataset of individual larval fish which are caught in a towed net.}
\usage{
simulate_catch_data(
  R = 1000,
  s0_av = 3,
  log10s0_sd = 0,
  g_av = 0.21,
  log10g_sd = 0,
  z_av = 0.25,
  log10z_sd = 0,
  max_age = 25,
  growth_in_weight = FALSE
)
}
\arguments{
\item{R}{Numebr of fish born per day}

\item{s0_av}{The mean size of individuals at hatch}

\item{log10s0_sd}{Variance of individuals around the mean size at hatch}

\item{g_av}{The mean growth rate of individuals in the cohort}

\item{log10g_sd}{Variance of individual growth rates around the mean growth rate of the cohort}

\item{z_av}{The mean mortality rate of individuals in the cohort}

\item{log10z_sd}{Variance of individual growth rates around the mean mortality rate of the cohort}

\item{growth_in_weight}{if TRUE growth rate is considered as the instanteous growth coefficient in weight (i.e  "size_sampled = size_birth * exp(growth_rate*age)"), If FALSE growth is modelled as change in length using a linear model.}
}
\value{
a tibble of simulated data
}
\description{
Simulating a dataset of individual larval fish which are caught in a towed net.
}
