% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R, R/simulate_population.R
\name{simulate_cumulative_mortality_model1}
\alias{simulate_cumulative_mortality_model1}
\alias{simulate_cumulative_mortality}
\title{Simulated integrated mortality rate (= cumulative hazard) from age birth to end for each individual under a particular size-distribution model}
\usage{
simulate_cumulative_mortality_model1(individual_data)

simulate_cumulative_mortality(individual_data, pars)
}
\arguments{
\item{individual_data}{A dataframe of individual data. Must include columns size_birth, size, age and paramaters of the size-distribution model.}

\item{pars}{pars list from which $model can be taken from}
}
\value{
A vector of integrated mortality rates for each individual. Lenght is the same as number of rows in the input data
}
\description{
Simulated integrated mortality rate (= cumulative hazard) from age birth to end for each individual under a particular size-distribution model
}
