% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R, R/simulate_population.R
\name{simulate_growth_model1}
\alias{simulate_growth_model1}
\alias{simulate_growth}
\title{Simulated integrated growth from time_birth to time_end for each individual under a particular size-distribution model}
\usage{
simulate_growth_model1(individual_data)

simulate_growth(individual_data, pars)
}
\arguments{
\item{individual_data}{A dataframe of individual data. Must include columns age and parameters of the size-distribution model.}

\item{model}{name of size-distribution model being modeled}
}
\value{
A vector of sizes for each individual. Length is the same as number of rows in the input data
}
\description{
Simulated integrated growth from time_birth to time_end for each individual under a particular size-distribution model
}
