% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model1.R, R/simulate_population.R
\name{size_dist_model1}
\alias{size_dist_model1}
\alias{size_dist_model}
\title{Predicted density distribution of individual sizes under a particular size-distribution model}
\usage{
size_dist_model1(x, pars)

size_dist_model(x, pars)
}
\arguments{
\item{x}{Size of individuals}

\item{pars}{list of parameter values for the size-distribution model, including model name}
}
\value{
Predicted size distribution for each x
}
\description{
Predicted density distribution of individual sizes under a particular size-distribution model
}
