% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizedist-package.R
\docType{package}
\name{sizedist-package}
\alias{sizedist-package}
\alias{sizedist}
\title{The 'sizedist' package.}
\description{
Fitting size distributions to estimate growth and mortality
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
