% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_bin_counts.R
\name{summarise_bin_counts}
\alias{summarise_bin_counts}
\title{Summarise counts for each bin interval}
\usage{
summarise_bin_counts(data, bin_var, bin_width)
}
\arguments{
\item{data}{A data frame or tibble}

\item{bin_var}{Numeric variable that bin intervals and counts are computed}

\item{bin_width}{The bin width for \code{age}bin_var`, usually the smallest resolution of re-sampling}
}
\value{
A tibble containing \code{bin_var}, lower and upper bounds and counts for each bin
}
\description{
Summarise counts for each bin interval
}
\examples{
Loblolly \%>\% summarise_bin_counts(height, bin_width = 0.5)

}
