% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bins.R
\name{add_bins}
\alias{add_bins}
\title{Creates bins}
\usage{
add_bins(data, var, bin_width)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. tibble) containing \code{age} and \code{size} variable as columns}

\item{var}{Name of variable, using tidy evaluation with data masking}

\item{bin_width}{The bin width for variable, usually the smallest resolution of resampling or measurement}
}
\value{
a data frame with binned variable added with suffix \verb{_bin}
}
\description{
add_bins() adds size and/or age bins to the dataset.
}
\examples{
\dontrun{

Loblolly \%>\% add_bins(var = height, bin_width = 0.5)
}

}
\seealso{
simulate_population()
}
