% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_pars.R
\name{add_pars}
\alias{add_pars}
\title{Join parameters object to Stan-composed data}
\usage{
add_pars(data, pars = NULL, type = NULL, prune = TRUE)
}
\arguments{
\item{data}{Binned data created by compose_count_data()}

\item{pars}{pars object}

\item{type}{Character string of model e.g. "model1"}

\item{prune}{If TRUE, parameters used for simulate_population() will be dropped}
}
\description{
Join parameters object to Stan-composed data
}
