% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pars.R
\name{add_pars}
\alias{add_pars}
\title{Add_pars Join_pars superseder}
\usage{
add_pars(data, pars = NULL, type = NULL, prune = TRUE)
}
\arguments{
\item{data}{Data created by summarise_bin_counts}

\item{pars}{pars object}

\item{type}{Character string of model}

\item{prune}{If TRUE, hyperparams used for simulate_population() will be dropped}
}
\value{

}
\description{
Add_pars Join_pars superseder
}
