% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_size_dist.R
\name{extract_fitted}
\alias{extract_fitted}
\title{Extract fitted values for Z, R, g}
\usage{
extract_fitted(fit, pars)
}
\arguments{
\item{fit}{model fit for growth model}

\item{pars}{pars list object containing hyperparameters}
}
\value{

}
\description{
Extract fitted values for Z, R, g
}
