% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dist.R
\name{fit_dist}
\alias{fit_dist}
\title{Fit size distribution model with Stan}
\usage{
fit_dist(stan_data)
}
\arguments{
\item{stan_data}{Stan-friendly data list created by XX function}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit size distribution model with Stan
}
