% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_pars.R
\name{join_stan_data}
\alias{join_stan_data}
\title{Merge bin counts and growth data for model3}
\usage{
join_stan_data(counts_list, growth_list)
}
\arguments{
\item{counts_list}{Stan-friendly count data created using \code{summarise_by_counts}}

\item{growth_list}{Stan-friendly growth data created using \code{compose_growth_data}}
}
\value{

}
\description{
Merge bin counts and growth data for model3
}
