% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{Simulates individuals sampled from population process with particular size-distribution model}
\usage{
simulate_population(
  time_end = 25,
  pars = default_pars("model1"),
  keep_dead = FALSE
)
}
\arguments{
\item{time_end}{Time to end simulation}

\item{pars}{list of parameter values for the size-distribution model, including model name}
}
\value{
a data frame (tibble) of simulated data
}
\description{
Simulates individuals sampled from population process with particular size-distribution model
}
